refactor : 사용하지 않는 메서드 , 변수 삭제. 패키지 구조 통일
