+++ src/main/java/com/lockers/outerpark/domain/reservation/service/ReservationServiceImpl.java
@@ -14,10 +14,10 @@ import com.lockers.outerpark.domain.reservation.dto.request.ReservationRequest;
 import com.lockers.outerpark.domain.reservation.dto.response.ReservationResponse;
 import com.lockers.outerpark.domain.reservation.dto.response.UserReservationResponse;
 import com.lockers.outerpark.domain.reservation.entity.Reservation;
-import com.lockers.outerpark.domain.reservation.entity.ReservationStatus;
 import com.lockers.outerpark.domain.reservation.exception.ReservationErrorCode;
 import com.lockers.outerpark.domain.reservation.exception.ReservationException;
 import com.lockers.outerpark.domain.reservation.repository.ReservationRepository;
+import com.lockers.outerpark.domain.reservation.type.ReservationStatus;
 import com.lockers.outerpark.domain.seat.entity.ReservationSeat;
 import com.lockers.outerpark.domain.seat.entity.Seat;
 import com.lockers.outerpark.domain.seat.service.SeatService;
