+++ src/main/java/com/lockers/outerpark/domain/reservation/type/ReservationStatus.java
@@ -1,4 +1,4 @@
-package com.lockers.outerpark.domain.reservation.entity;
+package com.lockers.outerpark.domain.reservation.type;
 
 public enum ReservationStatus {
 	CONFIRMED,
