+++ src/main/java/com/lockers/outerpark/domain/user/repository/UserRepository.java
@@ -7,9 +7,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.lockers.outerpark.domain.user.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    boolean existsByEmail(String email);
+	boolean existsByEmail(String email);
 
-    Optional<User> findByEmail(String email);
+	Optional<User> findByEmail(String email);
 
-    boolean existsByNickname(String nickname);
+	boolean existsByNickname(String nickname);
+
+	boolean existsByIdAndIsDeletedFalse(Long id);
 }
