+++ src/main/java/com/lockers/outerpark/domain/user/service/UserService.java
@@ -8,4 +8,6 @@ public interface UserService {
 	UserResponse getUserProfile(Long userId);
 
 	User getActiveUserById(Long userId);
+
+	void validateActiveUserById(Long userId);
 }
