+++ src/main/java/com/lockers/outerpark/domain/user/type/UserRole.java
@@ -1,0 +1,13 @@
+package com.lockers.outerpark.domain.user.type;
+
+public enum UserRole {
+	USER,
+	ADMIN;
+
+	// public static UserRole of(String role) {
+	//     return Arrays.stream(UserRole.values())
+	//             .filter(r -> r.name().equalsIgnoreCase(role))
+	//             .findFirst()
+	//             .orElseThrow(InvalidUserRoleException::new);
+	// }
+}
